Replace ARB_separate_shader_objects extension
This replaces the ARB_separate_shader_objects extension with traditional
linked shaders.  I was able to get the existing system to use linked
shaders without having to change any libobs graphics API.

This essentially creates a linked list of shader programs with
references to the shaders they link.  Before draw, it searches that
linked list for a particular pixel/vertex shader pair, and the linked
program associated with it.  If no matching program exists, it creates
the program.

+++ libobs-opengl/gl-shaderparser.c
@@ -617,7 +617,7 @@ static void gl_rename_attributes(struct gl_shader_parser *glsp)
 			val    = output_idx++;
 		}
 
-		dstr_printf(&new_name, "%s%u", prefix, val);
+		dstr_printf(&new_name, "%s%u", prefix, (unsigned int)val);
 		dstr_replace(&glsp->gl_string, attrib->name.array,
 				new_name.array);
 		dstr_move(&attrib->name, &new_name);
